Move libimage.ManifestList{Descriptor,Data} to libimage/define
These structs are (for better or worse) a part of
Podman's API, so podman-remote needs to include the subpackage that
defines them - which is all of libimage (and c/image/v5/copy) right now.

Instead, move them to libimage/define.

Alternatively, maybe Podman's API should not directly depend on (unstable)
c/common types?!

Only moves unchanged code, should not change behavior.
